utils: introduce heap_heap library
It tracks number of allocations and can help with detecting memory
leaks.

On Mac it was tested, on Linux was not, but most likely will work
the same.

The idea is that libc heap-related functions are expected to be
weak symbols. Meaning they can be redefined. The library simply
redefines the functions which students are expected to use and
tracks how many allocations remain unfreed.